# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "Siges/swagger.json",
        "version": "1.0"
    },
    "paths": {
        "/api/assessor/{assessor_id}": {
            "get": {
                "summary": "Visualiza os coordenadores com base no assessor_id.",
                "operationId": "d3850a0c9c4e6c6c5a9a2a0871d4eb16",
                "parameters": [
                    {
                        "name": "assessor_id",
                        "in": "path",
                        "description": "ID do assessor para buscar os coordenadores.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de vizualição de assessores"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Associa um usuário a uma coordenadoria (CDE) com base no ID do assessor.",
                "operationId": "482bc07b46b8ab72430efa23b7898255",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário a ser associado à coordenadoria.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "coordenadoria": {
                                        "type": "string",
                                        "example": "Coordenadoria Exemplo"
                                    },
                                    "municipio": {
                                        "type": "string",
                                        "example": "O Município é obrigatório apenas se a coordenadoria for regional"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário associado a uma coordenadoria com sucesso."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Adicione o token JWT no cabeçalho da requisição. Exemplo: 'Authorization: Bearer {token}'",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}